Now Using Parse and LQ. More coming
Refactor and update multiple files for improved functionality

This commit includes various updates and improvements across multiple files in the project. Key changes include:

- Removed unused namespaces and directives.
- Updated platform-specific initialization and data binding properties.
- Refactored methods for better clarity and functionality.
- Updated package references and project configurations.
- Enhanced user interface elements and debugging configurations.
- Improved code robustness and nullability handling.
- Added new functionalities such as live query setup and notification handling.
- Improved layout and functionality of the song details page.

+++ Dimmer/ViewModels/PlaylistSection.HomePageVM.cs
@@ -10,18 +10,18 @@ public partial class HomePageVM
     string? selectedPlaylistPageTitle;
 
     [ObservableProperty]
-    ObservableCollection<SongModelView> displayedSongsFromPlaylist=[];
+    ObservableCollection<SongModelView>? displayedSongsFromPlaylist=[];
     [ObservableProperty]
-    PlaylistModelView selectedPlaylistToOpenBtmSheet;
+    PlaylistModelView? selectedPlaylistToOpenBtmSheet = new();
     [ObservableProperty]
-    SongModelView selectedSongToOpenBtmSheet;
+    SongModelView? selectedSongToOpenBtmSheet=new();
     public void RefreshPlaylists()
     {
         DisplayedPlaylists?.Clear();
         DisplayedPlaylists = PlayBackService.GetAllPlaylists();
     }
     [ObservableProperty]
-    PlaylistModelView selectedPlaylist;
+    PlaylistModelView? selectedPlaylist=new();
 
     [RelayCommand]
     public async Task OpenSpecificPlaylistPage(string PlaylistID)//string playlistName)
@@ -68,9 +68,13 @@ public partial class HomePageVM
         else if (IsRemoveSong)
         {
             DisplayedSongsFromPlaylist.Remove(song);
-            PlayBackService.RemoveSongFromPlayListWithPlayListID(song, SelectedPlaylist.LocalDeviceId);
-            var toast = Toast.Make(songDeletedFromPlaylistText, duration);
-            await toast.Show(cts.Token);
+            SelectedPlaylist = DisplayedPlaylists.FirstOrDefault(x => x.LocalDeviceId == playlistModel.LocalDeviceId!);
+            if (SelectedPlaylist is not null)
+            {
+                PlayBackService.RemoveSongFromPlayListWithPlayListID(song, SelectedPlaylist.LocalDeviceId);
+                var toast = Toast.Make(songDeletedFromPlaylistText, duration);
+                await toast.Show(cts.Token);
+            }
         }
         else if (IsDeletePlaylist)
         {
